+++ src/pages/NoticeBoard/components/BoardList.tsx
@@ -1,11 +1,6 @@
 import React from 'react';
 import styled from 'styled-components';
-import dog1 from 'assets/images/dog1.jpg';
-import dog2 from 'assets/images/dog2.png';
-import dog3 from 'assets/images/dog3.png';
-import dog4 from 'assets/images/dog4.png';
-import dog5 from 'assets/images/dog5.png';
-import dog6 from 'assets/images/dog6.png';
+import BOARD_DATA from '../DATA/BOARD_DATA';
 
 const BoardList = () => {
   return (
@@ -29,57 +24,6 @@ const BoardList = () => {
     </BoardListWrapper>
   );
 };
-
-const BOARD_DATA = [
-  {
-    id: 1,
-    title: '우리 강아지',
-    date: '2020-01-01',
-    writer: 'choi',
-    views: 19000,
-    image: dog1,
-  },
-  {
-    id: 2,
-    title: '복덩이',
-    date: '2020-01-01',
-    writer: 'kim',
-    views: 4250,
-    image: dog2,
-  },
-  {
-    id: 3,
-    title: '신기하네요 글 길게쓰면?',
-    date: '2020-01-01',
-    writer: 'ahn',
-    views: 6500,
-    image: dog3,
-  },
-  {
-    id: 4,
-    title: '도배도배도배도배도배도배도배도배',
-    date: '2020-01-01',
-    writer: 'kang',
-    views: 200,
-    image: dog4,
-  },
-  {
-    id: 5,
-    title: '우리 강아지5',
-    date: '2020-01-01',
-    writer: 'yoon',
-    views: 200,
-    image: dog5,
-  },
-  {
-    id: 6,
-    title: '우리 강아지6',
-    date: '2020-01-01',
-    writer: 'hwang',
-    views: 200,
-    image: dog6,
-  },
-];
 
 const BoardListWrapper = styled.div`
   display: grid;
